% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_metadata}
\alias{create_metadata}
\title{Create Metadata Container}
\usage{
create_metadata()
}
\value{
A list with metadata and results.
}
\description{
Creates a list containing session metadata and an empty results object.
}
