% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DGM}
\alias{DGM}
\title{Data Generating Mechanism}
\usage{
DGM(
  N = 20000,
  autocorr_effects = 0.2,
  timepoints = 100,
  n_x_confounders = 5,
  n_y_confounders = 5,
  beta_focal = 0.1,
  beta_covfocal = 0.1,
  beta_covnofocal = 0.01,
  sigma = 0.5,
  seed = 427
)
}
\arguments{
\item{N}{Number of participants.}

\item{autocorr_effects}{Autoregressive effect size.}

\item{timepoints}{Number of timepoints.}

\item{n_x_confounders}{Number of X confounders.}

\item{n_y_confounders}{Number of Y confounders.}

\item{beta_focal}{Cross-lagged focal effect.}

\item{beta_covfocal}{Covariance among focal variables.}

\item{beta_covnofocal}{Covariance among non-focal variables.}

\item{sigma}{Residual variance.}

\item{seed}{Random seed.}
}
\value{
A list containing simulated data, parameters, and coefficient matrix.
}
\description{
Simulates data according to procedure by Bailey et al,2024.
}
