% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_clpm}
\alias{generate_clpm}
\title{Generate CLPM model specification}
\usage{
generate_clpm(T)
}
\arguments{
\item{T}{Integer. Number of timepoints (>= 2).}
}
\value{
A character string containing the lavaan model specification.
}
\description{
Generates a Cross-Lagged Panel Model (CLPM) specification in lavaan syntax.
}
