% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_riclpm}
\alias{generate_riclpm}
\title{Generate RI-CLPM model specification}
\usage{
generate_riclpm(T)
}
\arguments{
\item{T}{Integer. Number of timepoints (>= 3).}
}
\value{
A character string containing the lavaan model specification.
}
\description{
Generates a Random-Intercept Cross-Lagged Panel Model (RI-CLPM)
specification in lavaan syntax.
}
