% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{generate_segmented_riclpm}
\alias{generate_segmented_riclpm}
\title{Generate segmented RI-CLPM (MRI-CLPM) model specification}
\usage{
generate_segmented_riclpm(T, timespan = 3, across_seg_cov = "equal")
}
\arguments{
\item{T}{Integer. Total number of timepoints (>= 3).}

\item{timespan}{Integer. Length of each segment.}

\item{across_seg_cov}{Character. One of "equal", "zero", "free",
"toeplitz", or "toeplitz_cross".}
}
\value{
A character string containing the lavaan model specification.
}
\description{
Generates a segmented (multi–random-intercept) RI-CLPM specification.
}
