% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_crosslag_bias}
\alias{plot_crosslag_bias}
\title{Plot Cross-lagged Bias}
\usage{
plot_crosslag_bias(
  data,
  x_var,
  group_var,
  true_bx,
  true_by,
  x_label,
  group_label,
  title_suffix,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{Data frame containing coefficient lists.}

\item{x_var}{X-axis variable.}

\item{group_var}{Grouping variable.}

\item{true_bx}{True bx value.}

\item{true_by}{True by value.}

\item{x_label}{X-axis label.}

\item{group_label}{Legend label.}

\item{title_suffix}{Title suffix.}
}
\description{
Plots bias in cross-lagged parameters bx and by.
}
