% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_varRIx}
\alias{plot_varRIx}
\title{Plot Random Intercept Variance}
\usage{
plot_varRIx(
  data,
  x_var,
  group_var,
  x_label,
  group_label,
  plot_title,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{data}{Data frame.}

\item{x_var}{X-axis variable name.}

\item{group_var}{Grouping variable name.}

\item{x_label}{X-axis label.}

\item{group_label}{Legend label.}

\item{plot_title}{Plot title.}
}
\description{
Line plot of RI variance across conditions.
}
