% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_xy_sequences}
\alias{plot_xy_sequences}
\title{Plot X and Y Time Series}
\usage{
plot_xy_sequences(data, timepoints, n_participants = 5)
}
\arguments{
\item{data}{Matrix or data frame of simulated data.}

\item{timepoints}{Number of timepoints.}

\item{n_participants}{Number of participants to plot.}
}
\description{
Plots longitudinal X and Y trajectories for randomly sampled participants.
}
