% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{save_clear}
\alias{save_clear}
\title{Save Results and Clear Memory}
\usage{
save_clear(results, filename_prefix = "sim_results")
}
\arguments{
\item{results}{Object to save.}

\item{filename_prefix}{Prefix for the saved file.}
}
\value{
Filename of saved object.
}
\description{
Saves an R object to disk as an RDS file and triggers garbage collection.
}
